feat(adminSearchGroups): Move dropdown into separate component
