+++ src/app/features/admin/admin.module.ts
@@ -25,6 +25,7 @@ import { RegistrationRequestsComponent } from './registration-requests/registrat
 import { RejectedRegistrationRequestComponent } from './registration-requests/rejected-registration-request/rejected-registration-request.component';
 import { RejectedRegistrationRequestsComponent } from './registration-requests/rejected-registration-requests/rejected-registration-requests.component';
 import { ReportComponent } from './report/admin-report.component';
+import { EstablishmentDropdownComponent } from './search/establishment-dropdown/establishment-dropdown.component';
 import { SearchForGroupComponent } from './search/search-for-group/search-for-group.component';
 import { SearchForUserComponent } from './search/search-for-user/search-for-user.component';
 import { SearchForWorkplaceComponent } from './search/search-for-workplace/search-for-workplace.component';
@@ -49,6 +50,7 @@ import { SearchComponent } from './search/search.component';
     RejectedRegistrationRequestsComponent,
     RejectedRegistrationRequestComponent,
     ReportComponent,
+    EstablishmentDropdownComponent,
   ],
   providers: [
     LocalAuthoritiesReturnService,
