+++ src/app/features/admin/search/search-for-group/search-for-group.component.html
@@ -28,7 +28,7 @@
             <ng-container *ngFor="let item of results; index as workplaceIndex">
               <tr class="govuk-table__row">
                 <td class="govuk-table__cell">
-                  <a (click)="navigateToWorkplace(item.uid, '', item.nmdsId, $event)" href="#">{{ item.name }}</a>
+                  <a (click)="navigate(item.uid, '', item.nmdsId, $event)" href="#">{{ item.name }}</a>
                 </td>
                 <td class="govuk-table__cell">{{ item.nmdsId || '-' }}</td>
                 <td class="govuk-table__cell">
@@ -40,87 +40,12 @@
                   }}</a>
                 </td>
               </tr>
-              <tr
+              <app-establishment-dropdown
                 *ngIf="establishmentDetails[item.uid]"
-                class="govuk-panel--gray"
-                data-testid="groups-workplace-details"
-              >
-                <td colspan="6">
-                  <div class="govuk-!-margin-2">
-                    <table class="govuk-table">
-                      <thead class="govuk-table__head">
-                        <tr class="govuk-table__row">
-                          <th class="govuk-table__header govuk-!-width-one-third" scope="col">Address</th>
-                          <th class="govuk-table__header govuk-!-width-one-quarter" scope="col">Parent ID</th>
-                          <th class="govuk-table__header govuk-!-width-one-quarter" scope="col">Regulated</th>
-                          <th class="govuk-table__header" scope="col">Data Owner</th>
-                        </tr>
-                      </thead>
-                      <tbody class="govuk-table__body">
-                        <tr class="govuk-table__row govuk-util__vertical-align-top">
-                          <td class="govuk-table__cell govuk-!-font-weight-regular">
-                            {{ displayAddressForGroups(item) }}
-                          </td>
-                          <td class="govuk-table__cell govuk-!-font-weight-regular">
-                            <ng-container *ngIf="item.parent?.nmdsId; else noParentID">
-                              <a
-                                (click)="navigateToWorkplace(item.parent.uid, '', item.parent.nmdsId, $event)"
-                                href="#"
-                                >{{ item.parent.nmdsId }}</a
-                              >
-                            </ng-container>
-
-                            <ng-template #noParentID>-</ng-template>
-                          </td>
-                          <td class="govuk-table__cell govuk-!-font-weight-regular">
-                            {{ item.isRegulated ? 'Yes' : 'No' }}
-                          </td>
-                          <td colspan="2" class="govuk-table__cell govuk-!-font-weight-regular">
-                            {{ item.dataOwner }}
-                          </td>
-                        </tr>
-                      </tbody>
-                    </table>
-                  </div>
-                  <div class="govuk-!-margin-2" *ngIf="item.users?.length > 0">
-                    <table class="govuk-table">
-                      <thead class="govuk-table__head">
-                        <tr class="govuk-table__row">
-                          <th class="govuk-table__header govuk-!-width-one-third" scope="col">Users</th>
-                          <th class="govuk-table__header govuk-!-width-one-quarter" scope="col">Security question</th>
-                          <th class="govuk-table__header govuk-!-width-one-quarter" scope="col">Answer</th>
-                          <th class="govuk-table__header" scope="col">Locked</th>
-                        </tr>
-                      </thead>
-                      <tbody class="govuk-table__body">
-                        <tr
-                          class="govuk-table__row govuk-util__vertical-align-top"
-                          *ngFor="let user of item.users; index as userIndex"
-                        >
-                          <td class="govuk-table__cell govuk-!-font-weight-regular">
-                            {{ user.name }}
-                          </td>
-                          <td class="govuk-table__cell govuk-!-font-weight-regular">
-                            {{ user.securityQuestion }}
-                          </td>
-                          <td class="govuk-table__cell govuk-!-font-weight-regular">
-                            {{ user.securityAnswer }}
-                          </td>
-                          <td class="govuk-table__cell govuk-!-font-weight-regular">
-                            <a
-                              (click)="unlockWorkplaceUser(user.username, workplaceIndex, userIndex, $event)"
-                              href="#"
-                              *ngIf="user.isLocked; else notLocked"
-                              >Yes, unlock</a
-                            >
-                            <ng-template #notLocked>No</ng-template>
-                          </td>
-                        </tr>
-                      </tbody>
-                    </table>
-                  </div>
-                </td>
-              </tr>
+                [item]="item"
+                [workplaceIndex]="workplaceIndex"
+                (navigateToWorkplaceClicked)="navigateToWorkplace($event)"
+              ></app-establishment-dropdown>
             </ng-container>
           </tbody>
         </table>
