+++ src/app/features/admin/search/search-for-group/search-for-group.component.spec.ts
@@ -19,6 +19,7 @@ import { SharedModule } from '@shared/shared.module';
 import { fireEvent, render, within } from '@testing-library/angular';
 import { of } from 'rxjs';
 
+import { EstablishmentDropdownComponent } from '../establishment-dropdown/establishment-dropdown.component';
 import { SearchForGroupComponent } from './search-for-group.component';
 
 describe('SearchForGroupComponent', () => {
@@ -32,6 +33,7 @@ describe('SearchForGroupComponent', () => {
         FormsModule,
         ReactiveFormsModule,
       ],
+      declarations: [EstablishmentDropdownComponent],
       providers: [
         { provide: FeatureFlagsService, useClass: MockFeatureFlagsService },
         {
