+++ src/app/features/admin/search/search-for-group/search-for-group.component.ts
@@ -1,10 +1,8 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { EstablishmentSearchItem, GroupSearchRequest } from '@core/model/establishment.model';
-import { DialogService } from '@core/services/dialog.service';
 import { EstablishmentService } from '@core/services/establishment.service';
 import { SwitchWorkplaceService } from '@core/services/switch-workplace.service';
-import { AdminUnlockConfirmationDialogComponent } from '@shared/components/link-to-parent-cancel copy/admin-unlock-confirmation';
 import { Subscription } from 'rxjs';
 
 @Component({
@@ -23,7 +21,6 @@ export class SearchForGroupComponent implements OnInit, OnDestroy {
     private formBuilder: FormBuilder,
     private establishmentService: EstablishmentService,
     private switchWorkplaceService: SwitchWorkplaceService,
-    private dialogService: DialogService,
   ) {}
 
   ngOnInit(): void {
@@ -33,34 +30,19 @@ export class SearchForGroupComponent implements OnInit, OnDestroy {
     });
   }
 
-  public navigateToWorkplace(id: string, username: string, nmdsId: string, event: Event): void {
+  public navigate(id: string, username: string, nmdsId: string, event: Event): void {
     event.preventDefault();
-    this.switchWorkplaceService.navigateToWorkplace(id, username, nmdsId);
+    this.navigateToWorkplace({ id, username, nmdsId });
+  }
+
+  public navigateToWorkplace(workplace): void {
+    this.switchWorkplaceService.navigateToWorkplace(workplace.id, workplace.username, workplace.nmdsId);
   }
 
   public toggleDetails(uid: string, event: Event): void {
     event.preventDefault();
     this.establishmentDetails[uid] = !this.establishmentDetails[uid];
     this.establishmentDetailsLabel[uid] = this.establishmentDetailsLabel[uid] === 'Close' ? 'Open' : 'Close';
-  }
-
-  public displayAddressForGroups(workplace: EstablishmentSearchItem): string {
-    const secondaryAddress =
-      ' ' + [workplace.address2, workplace.town, workplace.county, workplace.postcode].filter(Boolean).join(', ') || '';
-
-    return workplace.address1 + secondaryAddress;
-  }
-
-  public unlockWorkplaceUser(username: string, workplaceIndex: number, userIndex: number, e: Event): void {
-    e.preventDefault();
-    const data = {
-      username,
-      removeUnlock: () => {
-        this.results[workplaceIndex].users[userIndex].isLocked = false;
-      },
-    };
-
-    this.dialogService.open(AdminUnlockConfirmationDialogComponent, data);
   }
 
   public onSubmit(): void {
